pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-document-extractor"
version = "0.1.0"
description = "End-to-end robust agentic document extractor using DPT architecture for Indian government documents"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "IPAC", email = "ipac@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.5.0",
    "pytesseract>=0.3.10",
    "opencv-python>=4.8.0",
    "transformers>=4.35.0",
    # layoutreader is optional - for advanced reading order prediction
    # Install from source if needed: pip install git+https://github.com/layoutreader/layoutreader.git
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "PyMuPDF>=1.23.0",
    "python-docx>=1.1.0",
    "Pillow>=10.0.0",
    "matplotlib>=3.8.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "pandas>=2.0.0",
    "click>=8.1.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
docextract = "agentic_document_extractor.cli.main:cli"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
